Added --standalone flag to --generate.
The --standalone flag controls whether a "common control body"
is generated in the runfile, so that it can be run on its own.
If --standalone is not specified, the common control body is
not generated, which allows the runfile to be included from
promises.cf or any other policy file.
The list of inputs necessary for the runfile to run is defined
as $(cfsketch_g.inputs), so that it can be referenced from
promises.cf as well.