Implemented interactive configuration mode.
Now the only mandatory parameter for the "define params" command is the
sketch name. If no file is given, the parameters will be queried
interactively. The interactive mode knows how to deal with scalars,
booleans, lists and arrays, and will do validation as values are
entered. If a sketch contains multiple bundles, it will also ask which
bundle to use.