Added require_order to Getopt::Long configuration
This makes it possible to execute cf-sketch commands with options from
the command line. For example, previously this:

$ cf-sketch.pl list -v

Would interpret the -v as the verbose flag for cf-sketch and not for
'list'. Adding require_order makes the cf-sketch option processing stop
when it encounters the first non-option argument, which gives us the
correct behavior.