Changed: restore and enhance colorization following Dan Barber's file renaming
Dan Barber had renamed a bunch of files (in organizing the materials
for publishing in book form) and this broke the colorization provided
by the "l.sh" tool.

I've updated the "remark" ruleset to:
1. colorize Title cards differently, and
2. indent non-Title files

which should make the TOC more readable -- it had grown into a real
"wall of text".

Aleksey